test(mcp-server): split the HTTP server suites and fix two weak assertions
The bounded-cache test asserted against a second verifier with its own
empty cache, so it re-tested negative caching and never exercised the
eviction it was named for. It now fills one verifier past maxEntries and
asserts that the oldest key has to be fetched again while the newest does
not. The cache-expiry test raced a 15ms sleep against a 5ms TTL and now
uses fake timers.

The concurrent-session test asserted only a 429, which any future limiter
would satisfy; it now asserts the session-cap body. Reading the SSE
session id waits for the endpoint frame rather than assuming it arrives in
the first chunk.

Both new test files were over the 300 line cap, so they are split by
concern into origin, session auth, and allocation suites over a shared
harness, which also removes the duplicated verifier stub and the repeated
loop-until-429 blocks.

The container no longer runs the network-reachable server as root, and an
IPv6 bind address is bracketed before being printed as a URL.

Documents that HTTP mode authenticates before routing, so the
documentation tools are not reachable without a key the way they are over
stdio.
